requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
fastapi==0.63.0
h11==0.12.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pydantic==1.7.3
pyparsing==2.4.6
pytoml==0.1.21
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
starlette==0.13.6
urllib3==1.25.8
uvicorn==0.13.3
webencodings==0.5.1
